// AMM/Izumi/types.ts
export type Token = {
    chainId: number
    address: string
    name: string
    symbol: string
    decimals: number
    logoURI: string
}

export type Pool = {
    tokenA: Token
    tokenB: Token
    pair: string
    fees: number
    tickSpacing: number,
    liquidity: bigint,
    sqrtPriceX96: bigint,
    tick: number,
    Quoter: Contract
    Pool: Contract
}
// ...
export type RemoveOptions = {
    slipage?: number
    deadline?: number
    percent?: number
    tokenId?: number
}

export type Position = {
    tokenId: number
    nonce: bigint
    operator: string
    token0: string
    token1: string
    fee: number
    tickLower: number
    tickUpper: number
    liquidity: bigint
    feeGrowthInside0LastX128: bigint
    feeGrowthInside1LastX128: bigint
    tokensOwed0: bigint
    tokensOwed1: bigint
}

export type Chains = 'arbitrum' | 'polygon' | 'optimism' | 'ethereum' | 'bsc' | 'polygonTestnet' | 'arbitrumTestnet' |
                    'ethereumTestnet' | 'optimismTestnet'